+++ CHANGELOG.md
@@ -1137,3 +1137,9 @@ Version 0.4.32 (15-04-2025)
 ---------------------------
 
 * Add hooks to expose dependencies for creating the node-signature.
+
+Version 0.4.33 (16-04-2025)
+---------------------------
+
+* Add support for subclass json serialisation / de-serialisation.
+* Make injectables an integer value to support migration to C++ in the NodeSignature